GUI: factorize common webpage code, avoid to duplicate too much
